Use separate table for storing annotations and transpose with crosstab.
This table (variant_annotation_table) stores each annotation name
and value, so it's flexible to add or remove. Then with crosstab,
rows are transposed into columns.
Although the overall performance of fetch_variants was not
penalized with crosstab, we finally dropped this idea because it
forced all annotations to have the same data type, as they were
all stored in the same column (annotation_value).